requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
databases==0.9.0
# alembic==1.11.0
python-dotenv==1.0.1
pydantic==2.10.3
psycopg2==2.9.10